add post

+++ src/app/components/add-post/add-post.component.ts
@@ -1,7 +1,9 @@
 import { Component, EventEmitter, OnInit, Output } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { Post } from 'src/app/models/post';
 import { AddPostService } from 'src/app/services/add-post.service';
+import { RouterServiceService } from 'src/app/services/router-service.service';
 
 
 @Component({
@@ -11,11 +13,12 @@ import { AddPostService } from 'src/app/services/add-post.service';
 })
 export class AddPostComponent implements OnInit {
 
-   post= {} as Post;
+  post= {} as Post;
   answer!: boolean;
-  @Output() answerEvent: EventEmitter<boolean>= new EventEmitter()
-  
-  constructor(private AddPostService: AddPostService, private readonly router: Router) { 
+  message: string ="";
+  imgFile?: File;
+  imgSrc=""
+  constructor(private AddPostService: AddPostService, private readonly router: Router,private dialog: MatDialog,private routerService: RouterServiceService) { 
 
   }
 
@@ -24,16 +27,47 @@ export class AddPostComponent implements OnInit {
   addpostdb(answer: boolean)
   {
       const newPost= {} as Post;
-      newPost.description=this.post.description;
-      newPost.imageSorce=this.post.imageSorce;
-        this.AddPostService.addNewPost(newPost).subscribe((result)=>{
-        console.log(result)
-        },(error)=>{
-          if(error.status==401){
-            this.router.navigate(['/login']);
-            localStorage.clear();
-          }
-        })
+      if(this.valid()){
+        newPost.description=this.post.description;
+        newPost.imageSorce=this.imgSrc;
+          this.AddPostService.addNewPost(newPost).subscribe((result)=>{
+            this.routerService.postChanged(true);
+         this.dialog.closeAll();
+
+          },(error)=>{
+            if(error.status==401){
+              this.router.navigate(['/login']);
+              localStorage.clear();
+              this.dialog.closeAll();
+            }
+            else{
+              this.message = error.error.detail
+            }
+          })
+      }
   }
- 
+  valid(){
+    let answer = true;
+    if(!this.post.description){
+      this.message="Enter description"
+      answer = false;
+    }
+    else if(!this.imgSrc){
+      this.message="Enter imagesorce"
+      answer=false
+    }
+    return answer
+  }
+  showFilePreview(event: any) {
+    console.log('here');
+    this.imgFile = event.target.files[0];
+    if (this.imgFile) {
+      var reader = new FileReader();
+      reader.onload = (e: any) => {
+        this.imgSrc = e.target.result;
+        console.log(this.imgSrc);
+      };
+      reader.readAsDataURL(this.imgFile);
+    }
+  }
 }
